# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "LDSGE API",
        "description": "LDSGE API",
        "contact": {
            "email": "luan@ldcamargo.com"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8000/",
    "basePath": "api/v1",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
...
        "/refresh": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Refresh Token",
                "description": "Refresh Token",
                "operationId": "refresh",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "default": []
                    }
                ]
            }
        },
...
    },
...
    "securityDefinitions": {
        "default": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}